requirements.txt
streamlit==1.40.2
pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
